Fix .gitignore rules

+++ main.py
@@ -3,19 +3,17 @@ from pydantic import BaseModel
 from model_class import CourseCompletionModel
 
 app = FastAPI()
-
 model = CourseCompletionModel()
 
 class StudentInput(BaseModel):
-    age: float
-    hours_per_week: float
-    assignments_submitted: float
+    age: int
+    hours_per_week: int
+    assignments_submitted: int
     desktop: int
     mobile: int
     pager: int
     smart_tv: int
     tablet: int
-        
 
 @app.get("/")
 def home():
@@ -25,7 +23,4 @@ def home():
 def predict_course(data: StudentInput):
     input_data = data.dict()
     result = model.predict(input_data)
-
-    return {
-        "prediction": "Completed" if result == 1 else "Not Completed"
-    }
+    return {"prediction": "Completed" if result == 1 else "Not Completed"}
